Update Front-end and backend scripts
CreateReportsTable -> renamed version of CreateTeamTable but also removed the "TeamNum" column since its in the studentinfo table

CreateStudentInfo Table -> updated it to have a primary key of s_id since there was no primary key specified

Updated InsertSampleRow -> matched it with the change in the format of the reports column and made the s_id match altamash's test.csv file data

StatusReportInput.php -> updated it to check if the secret ID entered corresponds to one in the DB or else it will notify the user that the code is invalid. also got rid of an extra echo statement that i forgot to delete which was displaying the timelog number

TAhomepage - i added a button to navigate to the upload csv page

StudentHomepage and TAhomepage-> i updated the sql query to display the data based on the changes made to the tables in the DB

Uploading_csv.php -> added a back button to navigate back to the homepage.